requirements.txt
imutils
numpy==1.21.0
opencv-python==4.5.3.56
pyzmq
pytest